Fix install directory of included files
