+++ src/serial.cpp
@@ -5,8 +5,8 @@
 #include <iostream>
 #include <string>
 
-#include "create/serial.h"
-#include "create/types.h"
+#include "libcreate/serial.h"
+#include "libcreate/types.h"
 
 namespace create
 {
