+++ src/serial_stream.cpp
@@ -4,8 +4,8 @@
 #include <iostream>
 #include <vector>
 
-#include "create/serial_stream.h"
-#include "create/types.h"
+#include "libcreate/serial_stream.h"
+#include "libcreate/types.h"
 
 namespace create
 {
