+++ tests/test_create.cpp
@@ -25,8 +25,8 @@ CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 POSSIBILITY OF SUCH DAMAGE.
 */
-#include "create/create.h"
-#include "create/types.h"
+#include "libcreate/create.h"
+#include "libcreate/types.h"
 
 #include "gtest/gtest.h"
