+++ tests/test_data.cpp
@@ -28,9 +28,9 @@ CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 POSSIBILITY OF SUCH DAMAGE.
 */
-#include "create/data.h"
-#include "create/packet.h"
-#include "create/types.h"
+#include "libcreate/data.h"
+#include "libcreate/packet.h"
+#include "libcreate/types.h"
 
 #include "gtest/gtest.h"
