+++ tests/test_packet.cpp
@@ -28,8 +28,8 @@ CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 POSSIBILITY OF SUCH DAMAGE.
 */
-#include "create/packet.h"
-#include "create/types.h"
+#include "libcreate/packet.h"
+#include "libcreate/types.h"
 
 #include <string>
